modified to only run 1 at a time

+++ Neural_Networks/NN_training_scripts/Data_Assembly_Scripts/get_above_below_probabilities_per_cluster.m
@@ -9,5 +9,5 @@ split_data = split(only_above_below_vars.',"_");
 sorted_thresholds = sort(str2double(split_data(:,end)));
 
 %create an array which will return the values
-threshold_probabilities = blind_pass_table{:,strcat("above_below_",string(sorted_thresholds))};
+threshold_probabilities = double(blind_pass_table{:,strcat("above_below_",string(sorted_thresholds))});
 end
